// src/app/products/dto/create-product.dto.ts
export interface CreateProductDto {
  productName: string;
  productCode: string;
  releaseDate: string;
  description: string;
  price: number;
  starRating: number;
  imageUrl: string;
  tags: string[];
}

export interface UpdateProductDto {
  productId: number;
  productName: string;
  productCode: string;
  releaseDate: string;
  description: string;
  price: number;
  starRating: number;
  imageUrl: string;
  tags: string[];
}